Endpoint for getting static content for a single rule
