+++ server/server.go
@@ -24,7 +24,6 @@ package server
 
 import (
 	"context"
-	"errors"
 	"io/ioutil"
 	"net/http"
 	"net/url"
@@ -245,26 +244,6 @@ func (server HTTPServer) proxyTo(baseURL string) func(http.ResponseWriter, *http
 			log.Error().Err(err).Msgf("Error writing the response")
 			handleServerError(writer, err)
 		}
-	}
-}
-
-// getGroups retrives the groups configuration from a channel to get the latest valid one and send the response back to the client
-func (server *HTTPServer) getGroups(writer http.ResponseWriter, request *http.Request) {
-	groupsConfig := <-server.GroupsChannel
-	if groupsConfig == nil {
-		err := errors.New("No groups retrieved")
-		log.Error().Err(err).Msg("Groups cannot be retrieved from content service. Check logs")
-		handleServerError(writer, err)
-		return
-	}
-
-	responseContent := make(map[string]interface{})
-	responseContent["status"] = "ok"
-	responseContent["groups"] = groupsConfig
-	err := responses.SendOK(writer, responseContent)
-	if err != nil {
-		log.Error().Err(err).Msg("Cannot send response")
-		handleServerError(writer, err)
 	}
 }
